Backend/requirements.txt
fastapi
uvicorn
motor
pydantic
pydantic-settings
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dotenv
email-validator